cleaned up routes/consolidated in main
